[SLIP-30] Add properties file for POST request payload (post.properties)
This commit introduces a new properties file, post.properties, which includes key-value pairs defining properties for the payload of a POST request to create a pet entity. The properties include "petId," "petName," "petCategory," "petPhotoUrls," "petTags," and "petStatus." This file is intended for use in API tests where specific pet data needs to be configured and provided dynamically during the creation of a pet.